moving ip address to settings popup
